pyproject.toml
[project]
name = "ocrean"
version = "0.1.0"
description = ""
authors = [
    {name = "maxence"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    "pypdf>=4.0.0",
    "pdf2image>=1.16.0",
    "kss>=3.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "ruff (>=0.14.6,<0.15.0)",
    "pre-commit (>=4.4.0,<5.0.0)",
    "pytest (>=8.0.0,<9.0.0)"
]
